+++ src/trackers/simple/SimpleTrackerModule.ts
@@ -933,6 +933,10 @@ export class SimpleTrackerModule implements TrackerModule<SimpleTrackerEntry, Si
 			e.stopPropagation();
 			e.stopImmediatePropagation();
 		});
+		el.addEventListener('mouseup', (e) => {
+			e.stopPropagation();
+			e.stopImmediatePropagation();
+		});
 		el.addEventListener('pointerup', (e) => {
 			e.preventDefault();
 			e.stopPropagation();