+++ src/widget/EventHistorySection.ts
@@ -203,6 +203,10 @@ export class EventHistorySection {
 			e.stopPropagation();
 			e.stopImmediatePropagation();
 		});
+		el.addEventListener('mouseup', (e) => {
+			e.stopPropagation();
+			e.stopImmediatePropagation();
+		});
 		el.addEventListener('pointerup', (e) => {
 			e.preventDefault();
 			e.stopPropagation();