+++ src/widget/QuickActions.ts
@@ -189,6 +189,10 @@ export class QuickActions {
 			e.stopPropagation();
 			e.stopImmediatePropagation();
 		});
+		el.addEventListener('mouseup', (e) => {
+			e.stopPropagation();
+			e.stopImmediatePropagation();
+		});
 		el.addEventListener('pointerup', (e) => {
 			e.preventDefault();
 			e.stopPropagation();