+++ src/widget/shared/InlineEditPanel.ts
@@ -106,6 +106,10 @@ export class InlineEditPanel {
 			e.stopPropagation();
 			e.stopImmediatePropagation();
 		});
+		el.addEventListener('mouseup', (e) => {
+			e.stopPropagation();
+			e.stopImmediatePropagation();
+		});
 		el.addEventListener('pointerup', (e) => {
 			e.preventDefault();
 			e.stopPropagation();